Fix login test url rewriting & refactor
