+++ aidants_connect_web/tests/test_functional/test_view_mandats.py
@@ -1,9 +1,10 @@
 from django.contrib.staticfiles.testing import StaticLiveServerTestCase
-from django.core import mail
 
 from django.test import tag
 from selenium.webdriver.firefox.webdriver import WebDriver
 from aidants_connect_web.models import Aidant, Usager, Mandat
+from aidants_connect_web.tests.test_functional.utilities import login_aidant
+
 from datetime import date
 
 
@@ -72,7 +73,7 @@ class ViewMandats(StaticLiveServerTestCase):
         self.selenium.get(f"{self.live_server_url}/dashboard/")
 
         # Login
-        self.login_aidant()
+        login_aidant(self)
 
         # Dashboard
         self.selenium.find_element_by_id("view_mandats").click()
@@ -83,19 +84,3 @@ class ViewMandats(StaticLiveServerTestCase):
         # The first row is the title row
         mandats = rows[1:]
         self.assertEqual(len(mandats), 3)
-
-    def login_aidant(self):
-        login_field = self.selenium.find_element_by_id("id_email")
-        login_field.send_keys("thierry@thierry.com")
-        submit_button = self.selenium.find_element_by_xpath("//button")
-        submit_button.click()
-        email_sent_title = self.selenium.find_element_by_tag_name("h1").text
-        self.assertEqual(
-            email_sent_title, "Un email vous a été envoyé pour vous connecter."
-        )
-        self.assertEqual(len(mail.outbox), 1)
-        token_email = mail.outbox[0].body
-        line_containing_magic_link = token_email.split("\n")[2]
-        magic_link_https = line_containing_magic_link.split()[-1]
-        magic_link_http = magic_link_https.replace("https", "http")
-        self.selenium.get(magic_link_http)
